// src/lib/models/SensorMeasurementsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SensorMeasurementsResponse = {
    id: string;
    ph: (number | null);
    azote: (number | null);
    phosphore: (number | null);
    potassium: (number | null);
    humidity: (number | null);
    temperature: (number | null);
    capteur_id: string;
    timestamp: string;
    measurements: Record<string, any>;
    parcelle_id: string;
    created_at: string;
    updated_at: string;
};